requirements.txt
flask~=2.3.2
flask_sqlalchemy
psycopg2-binary
PyJWT~=2.7.0
SQLAlchemy==2.0.17
Werkzeug==2.3.6


jsonschema~=4.17.3
pika~=1.3.2
requests~=2.31.0
python-dateutil~=2.8.2
python-crontab~=2.7.1
APScheduler~=3.10.1
schedule~=1.2.0